+++ CHANGES.rst
@@ -7,6 +7,8 @@ New Features
 Bug Fixes
 ^^^^^^^^^
 
+- Fix bug in fitting with weights if weights argument is set to 'unc'. [#979]
+
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^